Use converters over datatypes for variety.
The purge command wouldn't show how many messages were purged because of previously done imperfect practice, via embedding stat sender in a for-loop which would mostly not trigger after first.

The movemessage command generically put all attachments in a field even if it were embedable but if the first attachment was an renderable image, it puts in the embed, for rest or first one not being embedable it puts in "Attachments" field.

Some optimizations to the "movemessage bulk" command, removed redundant for-loop.